Add: tests for `GCloud.go!`

+++ spec/support/config.rb
@@ -7,14 +7,16 @@ module RSpec
 end
 
 RSpec.configure do |config|
-  config.before(:each) do
+  config.before(:each) do |test|
     stub_const('::Turbulence::Config::CONFIG_FILE', RSpec::Turbulence::TEST_CONFIG_FILE)
 
-    sample_contents = File.read("#{RSpec::Turbulence::TEST_CONFIG_FILE}.example")
-    File.write(RSpec::Turbulence::TEST_CONFIG_FILE, sample_contents)
+    unless test.metadata[:empty_config]
+      sample_contents = File.read("#{RSpec::Turbulence::TEST_CONFIG_FILE}.example")
+      File.write(RSpec::Turbulence::TEST_CONFIG_FILE, sample_contents)
+    end
   end
 
   config.after(:all) do
-    File.delete(RSpec::Turbulence::TEST_CONFIG_FILE)
+    File.delete(RSpec::Turbulence::TEST_CONFIG_FILE) if File.exist?(RSpec::Turbulence::TEST_CONFIG_FILE)
   end
 end
